refactor: change code for less conflicts during merge
