--- pointing-poker/src/components/main-content/MainContent.scss
@@ -1,66 +1,0 @@
-:root {
-  --color-dark: #2f2f2f;
-  --width-forms: 350px;
-}
-
-*,
-*::before,
-*::after {
-  box-sizing: inherit;
-  margin: 0;
-  padding: 0;
-}
-
-html {
-  box-sizing: border-box;
-  height: 100%;
-  font-size: 10px;
-}
-
-body {
-  background: var(--color-dark);
-  // display: flex;
-  // align-items: center;
-  // justify-content: center;
-  min-height: 100vh;
-  min-width: 100%;
-}
-
-#root {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  
-  header {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #0f5f5f;
-  }
-
-  main {
-    width: 80%;
-  }
-}
-
-.router__fade-enter {
-  opacity: 0;
-  z-index: 1;
-}
-
-.router__fade-enter.router__fade-enter-active {
-  opacity: 1;
-  transition: opacity 500ms ease-in;
-}
-
-.router__fade-exit {
-  opacity: 1;
-}
-
-.router__fade-exit.router__fade-exit-active {
-  opacity: 0;
-  transition: opacity 500ms ease-in;
-}
